Move CreateFrame into seperate file
